+++ lib/rubocop/cop/lint/literal_in_condition.rb
@@ -3,8 +3,20 @@
 module Rubocop
   module Cop
     module Lint
-      # This cop checks for literals used as operands in and/or
-      # expressions in the conditions of if/while/until.
+      # This cop checks for literals used as the conditions or as
+      # operands in and/or expressions serving as the conditions of
+      # if/while/until.
+      #
+      # @example
+      #
+      #   if 20
+      #     do_something
+      #   end
+      #
+      #   if some_var && true
+      #     do_something
+      #   end
+      #
       class LiteralInCondition < Cop
         MSG = 'Literal %s appeared in a condition.'
 
@@ -46,8 +58,14 @@ module Rubocop
         def check_for_literal(node)
           cond, = *node
 
-          on_node([:and, :or], cond) do |logic_node|
-            *operands = *logic_node
+          # if the code node is literal we obviously have a problem
+          if LITERALS.include?(cond.type)
+            add_offence(:warning, cond.loc.expression,
+                        format(MSG, cond.loc.expression.source))
+          elsif [:and, :or].include?(cond.type)
+            # alternatively we have to consider a logical node with a
+            # literal argument
+            *operands = *cond
             operands.each do |op|
               if LITERALS.include?(op.type)
                 add_offence(:warning, op.loc.expression,