+++ lib/rubocop/cop/lint/end_alignment.rb
@@ -5,23 +5,19 @@ module Rubocop
     module Lint
       # This cop checks whether the end keywords are aligned properly.
       #
-      # There are two ways to align the end of a block.
-      # Set BlockAlignSchema to StartOfAssignment if you want to
-      # align the end to the beginning of assignment expression.
-      # This is the default behavior.
+      # For keywords (if, def, etc.) the end is aligned with the start
+      # of the keyword.
+      # For blocks - with the start of the expression where the block
+      # is defined.
+      #
       # @example
+      #
+      #   variable = if true
+      #              end
       #
       #   variable = lambda do |i|
       #     i
       #   end
-      #
-      # Set BlockAlignSchema to StartOfBlockCommand if you want to
-      # align the end to the beginning of the expression that called the block.
-      # @example
-      #
-      #   variable = lambda do |i|
-      #                i
-      #              end
       class EndAlignment < Cop
         MSG = 'end at %d, %d is not aligned with %s at %d, %d'
 
@@ -79,6 +75,8 @@ module Rubocop
         end
 
         def on_and(node)
+          return if already_processed_node?(node)
+
           _left, right = *node
           if right.type == :block
             check_block_alignment(node.loc.expression, right.loc)
@@ -90,10 +88,8 @@ module Rubocop
         alias_method :on_or, :on_and
 
         def on_lvasgn(node)
-          if align_with_start_of_assignment?
-            _, children = *node
-            process_block_assignment(node, children)
-          end
+          _, children = *node
+          process_block_assignment(node, children)
           super
         end
 
@@ -104,57 +100,48 @@ module Rubocop
         alias_method :on_or_asgn,  :on_lvasgn
 
         def on_casgn(node)
-          if align_with_start_of_assignment?
-            _, _, children = *node
-            process_block_assignment(node, children)
-          end
+          _, _, children = *node
+          process_block_assignment(node, children)
           super
         end
 
         def on_op_asgn(node)
-          if align_with_start_of_assignment?
-            variable, _op, args = *node
-            process_block_assignment(variable, args)
-          end
+          variable, _op, args = *node
+          process_block_assignment(variable, args)
           super
         end
 
         def on_send(node)
-          if align_with_start_of_assignment?
-            receiver, method, args = *node
-            if attribute_writer?(method)
-              process_block_assignment(receiver, args)
-            end
-          end
+          _receiver, _method, *args = *node
+          process_block_assignment(node, args.last)
           super
         end
 
         def on_masgn(node)
-          if align_with_start_of_assignment?
-            variables, args = *node
-            process_block_assignment(variables, args)
-          end
+          variables, args = *node
+          process_block_assignment(variables, args)
           super
         end
 
         private
 
         def process_block_assignment(begin_node, block_node)
-          if block_node
-            while block_node.type == :send
-              receiver, _method, args = *block_node
-              if receiver && [:block, :send].include?(receiver.type)
-                block_node = receiver
-              elsif args && [:block, :send].include?(args.type)
-                block_node = args
-              else
-                break
-              end
+          return unless block_node
+          return if already_processed_node?(block_node)
+
+          while block_node.type == :send
+            receiver, _method, args = *block_node
+            if receiver && [:block, :send].include?(receiver.type)
+              block_node = receiver
+            elsif args && [:block, :send].include?(args.type)
+              block_node = args
+            else
+              break
             end
-            if block_node.type == :block
-              @inspected_blocks << block_node
-              check_block_alignment(begin_node.loc.expression, block_node.loc)
-            end
+          end
+          if block_node.type == :block
+            @inspected_blocks << block_node
+            check_block_alignment(begin_node.loc.expression, block_node.loc)
           end
         end
 
@@ -187,14 +174,6 @@ module Rubocop
 
         def already_processed_node?(node)
           @inspected_blocks.include?(node)
-        end
-
-        def attribute_writer?(method)
-          method.to_s[-1] == '='
-        end
-
-        def align_with_start_of_assignment?
-          EndAlignment.config['BlockAlignSchema'] == 'StartOfAssignment'
         end
       end
     end