+++ lib/rubocop/cop/lint/ensure_return.rb
@@ -12,7 +12,7 @@ module Rubocop
 
           on_node(:return, ensure_body) do |e|
             add_offence(:warning, e.loc.expression, MSG)
-          end
+          end if ensure_body
 
           super
         end