+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#318](https://github.com/bbatsov/rubocop/issues/318) - correct some special cases of block end alignment
 * [#317](https://github.com/bbatsov/rubocop/issues/317) - fix a false positive in `LiteralInCondition`
+* [#317](https://github.com/bbatsov/rubocop/issues/322) - Fix exception of `UnusedLocalVariable` and `ShadowingOuterLocalVariable` when inspecting keyword splat argument
 
 ## 0.9.0 (01/07/2013)